+++ R/decoration.R
@@ -50,6 +50,18 @@
 #'   present, replaces the existing one automatically: e.g., adding
 #'   \code{"peak.labels"} replaces\code{"peaks"} if present.
 #'
+#'   The annotation layers are added to the plot using statistics defined in 'ggspectra':
+#'   \code{\link{stat_peaks}}, \code{\link{stat_valleys}},
+#'   \code{\link{stat_label_peaks}}, \code{\link{stat_label_valleys}},
+#'   \code{\link{stat_find_wls}}, \code{\link{stat_spikes}},
+#'   \code{\link{stat_wb_total}}, \code{\link{stat_wb_mean}},
+#'   \code{\link{stat_wb_irrad}}, \code{\link{stat_wb_sirrad}},
+#'   \code{\link{stat_wb_contribution}}, \code{\link{stat_wb_relative}},
+#'   and \code{\link{stat_wl_strip}}. However, only some of their parameters
+#'   can be passed arguments through \code{autoplot} methods. In some cases
+#'   the defaults used by \code{autoplot} methods are not the defaults of the
+#'   statistics.
+#'
 #' @details Vectors of character strings passed as argument to
 #' \code{annotations} are parsed so that if the first member string is
 #' \code{"+"}, the remaining members are added to the current default for
@@ -77,7 +89,7 @@ decoration <- function(w.band,
                        y.expanse = y.max - y.min,
                        annotations,
                        span,
-                       strict = is.null(span),
+                       strict = FALSE,
                        wls.target = "HM",
                        label.qty,
                        label.mult = 1,